core/pyproject.toml
[project]
name = "remoteagent-core"
version = "0.1.0"
description = "Shared core for remoteAgent bots (Feishu, sessions, streaming)"
requires-python = ">=3.10"
dependencies = [
    "lark-oapi>=1.4.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
